[DAT-8479] Move addPrefixToExceptionMessage into DefaultInputHandler and delete UiIllegalArgumentExceptionMessageHandler
